requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
bracex==2.6
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dataclasses-json==0.6.7
decorator==5.2.1
deepagents==0.2.0
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
fastapi==0.121.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-anthropic==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-deepseek==1.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
loguru==0.7.3
MarkupSafe==3.0.3
marshmallow==3.26.1
minio==7.2.10
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.49.3
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcmatch==10.1
wcwidth==0.2.14
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
